requirements.txt
torch==2.7.0
torchvision==0.22.0
torchaudio==2.7.0
transformers==4.45.2
huggingface-hub==0.35.2
accelerate==1.7.0
bitsandbytes==0.46.0
safetensors==0.5.3
tokenizers==0.20.3
tqdm==4.67.1
pillow==11.1.0
scipy==1.15.1
requests==2.32.4
pyyaml==6.0.2
peft==0.15.2
timm==1.0.15
einops==0.8.1
wheel
flash_attn
sentencepiece==0.2.0
aiohttp
bs4
lxml